+++ tests/test_processes.py
@@ -1,6 +1,7 @@
 import re
 
 import pytest
+import dirty_equals
 
 from openeo_driver.datacube import DriverDataCube
 from openeo_driver.errors import (
@@ -221,6 +222,43 @@ def test_process_registry_add_function_argument_names():
     assert 'largest value' in spec['description']
     assert all(k in spec for k in ['parameters', 'returns'])
     assert reg.get_function('max') is max
+
+
+def test_process_registry_add_function_no_override():
+    reg = ProcessRegistry()
+
+    reg.add_spec_by_name("min", "max")
+
+    with pytest.raises(ProcessRegistryException, match="Process 'max' already defined in namespace 'backend'"):
+
+        @reg.add_function
+        def max(*args):
+            return max(*args)
+
+    assert reg.contains("max")
+
+
+def test_process_registry_add_function_allow_override():
+    reg = ProcessRegistry()
+
+    reg.add_spec_by_name("min", "max")
+
+    assert reg.get_spec("max") == dirty_equals.IsPartialDict(
+        {
+            "id": "max",
+            "summary": "Maximum value",
+        }
+    )
+
+    @reg.add_function(allow_override=True, spec={"id": "max", "summary": "Ultra pro max!"})
+    def max(*args):
+        return max(*args)
+
+    assert reg.contains("max")
+    assert reg.get_spec("max") == {
+        "id": "max",
+        "summary": "Ultra pro max!",
+    }
 
 
 def test_process_registry_with_spec_040():